Separate color commands into cog
