+++ bot/extensions/fun/fun_cog.py
@@ -1,9 +1,7 @@
-import os
 from json import loads
-from PIL import Image
 from discord.ext.commands.cooldowns import BucketType
 from discord.ext.commands import Cog, command, cooldown
-from discord import Embed, File
+from discord import Embed
 from requests import get
 import random
 from bot.models.extensions.fun.eightball.answer import Answer
@@ -41,35 +39,6 @@ class FunCog(Cog, name="Fun", description="Collection of fun commands"):
         )
 
         await ctx.send(embed=embed)
-
-    async def display_color(self, color_img, ctx):
-        color_img.save('color.png')
-
-        embed = Embed(
-            color=self.bot.default_color,
-            description='Here goes your color!'
-        )
-        file = File('color.png')
-        await ctx.send(embed=embed, file=file)
-
-        os.remove('color.png')
-    
-
-    @command(name='rgb', help='''Displays the RGB color entered by the user.''')
-    async def rgb_command(self, ctx, r: int, g: int, b: int):
-        img = Image.new('RGB', (200, 200), (r, g, b))
-
-        await self.display_color(img, ctx)
-
-
-    @command(name='hex', help='''Displays the color of the hexcode entered by the user.''')
-    async def hex_command(self, ctx, hex: str):
-        if not hex.startswith('#'):
-            hex = '#' + hex
-
-        img = Image.new('RGB', (200, 200), hex)
-
-        await self.display_color(img, ctx)
